Event Selection Menu
Added a menu to select which GW Event to display. Also added a way to read a text file containing information related to the same GW Event, so that in future versions of the software these info can be displayed.
As with the previous version, it is necessary to manually copy files from the "copyme" folder into the appropriate GWSky application folder.